# docs/swagger.yaml
definitions:
...
  models.Activities:
    properties:
      gambar:
        type: string
      id:
        type: integer
      image_url:
        type: string
      tanggal:
        type: string
      title:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /activities:
    get:
      description: Retrieves all Activities from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Activities'
            type: array
      summary: Get All Activities
      tags:
      - Activities
    post:
      consumes:
      - multipart/form-data
      description: Uploads a Activities along with its file and saves them to the
        database.
      parameters:
      - description: Title Activities
        in: formData
        name: title
        required: true
        type: string
      - description: Tanggal Activities
        in: formData
        name: tanggal
        required: true
        type: string
      - description: File gambar
        in: formData
        name: gambar
        required: true
        type: file
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Activities'
      summary: Upload an activity with File
      tags:
      - Activities
  /activities/{id}:
    delete:
      description: Deletes a Activity and its gambar from the database and storage.
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Activity deleted successfully
          schema:
            type: string
      summary: Delete Activity
      tags:
      - Activities
    get:
      description: Retrieves Activity data by its ID.
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get Activity By ID
      tags:
      - Activities
    put:
      consumes:
      - multipart/form-data
      description: Edits a Activity including the ability to replace its file.
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Title Activity
        in: formData
        name: title
        required: true
        type: string
      - description: Tanggal Activity
        in: formData
        name: tanggal
        required: true
        type: string
      - description: Gambar Activity (optional)
        in: formData
        name: gambar
        type: file
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Activities'
      summary: Edit Activity
      tags:
      - Activities
  /activities/file/{id}:
    get:
      description: Retrieves the image of a activity by its ID.
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get Gambar Activity
      tags:
      - Activities
...
swagger: "2.0"